Bump deps and use npm7
