--- .stylelintignore
@@ -1,1 +1,0 @@
-/src/fonts
