+++ babel.config.js
@@ -1,11 +1,3 @@
-module.exports = {
-	presets: [
-		[
-			'@babel/preset-env',
-			{
-				useBuiltIns: 'usage',
-				corejs: 3
-			}
-		]
-	]
-};
+const babelConfig = require('@nextcloud/babel-config')
+
+module.exports = babelConfig
