+++ package.json
@@ -24,9 +24,9 @@
 	"license": "agpl",
 	"private": true,
 	"scripts": {
-		"dev": "webpack --config webpack.dev.js",
-		"watch": "webpack --progress --watch --config webpack.dev.js",
-		"build": "webpack --progress --config webpack.prod.js",
+		"build": "NODE_ENV=production webpack --progress --config webpack.js",
+		"dev": "NODE_ENV=development webpack --progress --config webpack.js",
+		"watch": "NODE_ENV=development webpack --progress --watch --config webpack.js",
 		"lint": "eslint --ext .js,.vue src",
 		"lint:fix": "eslint --ext .js,.vue src --fix",
 		"stylelint": "stylelint src css",
@@ -83,48 +83,21 @@
 		"npm": ">=7.0.0"
 	},
 	"devDependencies": {
-		"@babel/core": "^7.14.6",
-		"@babel/plugin-syntax-dynamic-import": "^7.8.3",
-		"@babel/preset-env": "^7.14.7",
-		"@nextcloud/browserslist-config": "^1.0.0",
-		"@nextcloud/eslint-config": "^1.2.0",
-		"@nextcloud/eslint-plugin": "^1.5.0",
+		"@nextcloud/babel-config": "^1.0.0-beta.1",
+		"@nextcloud/browserslist-config": "^2.1.0",
+		"@nextcloud/eslint-config": "^5.1.0",
+		"@nextcloud/stylelint-config": "^1.0.0-beta.0",
+		"@nextcloud/webpack-vue-config": "^4.0.3",
 		"@vue/test-utils": "^1.2.1",
 		"babel-core": "^7.0.0-bridge.0",
-		"babel-eslint": "^10.0.3",
 		"babel-jest": "^27.0.5",
-		"babel-loader": "^8.2.2",
-		"css-loader": "^3.6.0",
-		"eslint": "^5.16.0",
-		"eslint-config-standard": "^12.0.0",
-		"eslint-friendly-formatter": "^4.0.1",
-		"eslint-loader": "^3.0.3",
-		"eslint-plugin-import": "^2.23.4",
-		"eslint-plugin-node": "^11.1.0",
-		"eslint-plugin-promise": "^4.3.1",
-		"eslint-plugin-standard": "^4.1.0",
-		"eslint-plugin-vue": "^5.2.3",
-		"file-loader": "^6.2.0",
+		"babel-loader-exclude-node-modules-except": "^1.2.1",
 		"iconfont-plugin-webpack": "^1.1.4",
 		"jest": "^27.0.5",
 		"jest-serializer-vue": "^2.0.2",
-		"prettier-eslint": "^12.0.0",
-		"regenerator-runtime": "^0.13.7",
 		"resolve-url-loader": "^4.0.0",
-		"sass": "^1.35.1",
-		"sass-loader": "^10.1.1",
-		"stylelint": "^13.13.1",
-		"stylelint-bare-webpack-plugin": "^2.1.0",
-		"stylelint-config-recommended-scss": "^4.2.0",
-		"stylelint-scss": "^3.19.0",
-		"terser-webpack-plugin": "^4.2.3",
-		"url-loader": "^4.1.1",
 		"vue-jest": "^3.0.7",
-		"vue-loader": "^15.9.7",
-		"vue-template-compiler": "^2.6.12",
-		"webpack": "^4.46.0",
-		"webpack-cli": "^4.7.2",
-		"webpack-merge": "^5.8.0"
+		"vue-template-compiler": "^2.6.12"
 	},
 	"optionalDependencies": {
 		"fsevents": "^2.3.2"
