+++ stylelint.config.js
@@ -1,0 +1,3 @@
+const stylelintConfig = require('@nextcloud/stylelint-config')
+
+module.exports = stylelintConfig
