+++ templates/main.php
@@ -19,5 +19,5 @@
  * License along with this library.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-script('calendar', 'calendar');
+script('calendar', 'calendar-main');
 style('calendar', 'calendar');
