--- webpack.prod.js
@@ -1,18 +1,0 @@
-const { merge } = require('webpack-merge')
-const common = require('./webpack.common.js')
-const TerserPlugin = require('terser-webpack-plugin')
-
-module.exports = merge(common, {
-	mode: 'production',
-	devtool: '#source-map',
-	optimization: {
-		minimizer: [new TerserPlugin({
-			terserOptions: {
-				output: {
-					comments: false,
-				}
-			},
-			sourceMap: true,
-		})],
-	}
-})
